requirements.txt
asgiref==3.5.2
autopep8==2.0.1
Django==3.2
djangorestframework==3.12.4
psycopg2==2.9.5
pycodestyle==2.10.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1